requirements.txt
discord.py==2.3.2
tweety-ns==2.0.8
python-dotenv==1.0.0
pyyaml==6.0.1
aiosqlite==0.18.0
google-genai>=0.1.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0